More tidying

+++ src/Vulkan/Queue.cpp
@@ -1,3 +1,28 @@
+////////////////////////////////////////////////////////////////////////////////
+// MIT License
+//
+// Copyright (c) 2021.  Shane Hyde (shane@noctonyx.com)
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+//
+////////////////////////////////////////////////////////////////////////////////
+
 //
 // Created by shane on 28/12/2020.
 //
@@ -66,12 +91,12 @@ namespace RxCore
 
         VkSubmitInfo si{};
         si.sType = VK_STRUCTURE_TYPE_SUBMIT_INFO;
-        si.waitSemaphoreCount = waitSems.size();
+        si.waitSemaphoreCount = static_cast<uint32_t>(waitSems.size());
         si.pWaitSemaphores = waitSems.data();
         si.pWaitDstStageMask = waitStages.data();
-        si.commandBufferCount = buffer_handles.size();
+        si.commandBufferCount = static_cast<uint32_t>(buffer_handles.size());
         si.pCommandBuffers = buffer_handles.data();
-        si.signalSemaphoreCount = signalSemaphores.size();
+        si.signalSemaphoreCount = static_cast<uint32_t>(signalSemaphores.size());
         si.pSignalSemaphores = signalSemaphores.data();
 
         //VkSubmitInfo si{waitSems, waitStages, buffer_handles, signalSemaphores};
